code refactored and splitted in modules
