+++ blockly_server/app/assets/code_templates/admin_parameters.yaml
