+++ blockly_server/app/assets/code_templates/template.xml
