+++ blockly_server/app/assets/sound_effects/machine_gun.mp3
