+++ blockly_server/app/assets/systray_mode.py
